MT#19043 Check perms 0755 on /etc/ngcp-config/ssl for PRO only
For some unknown for me reason they are different on CE and PRO.
From the installer code they should be 0750,
while something changes them to 0755 for PRO only.

I spend two hours trying to find the reason here - no luck.
Updating the test to fix tests with the actual state and
continue with upgrade fixes here.

Change-Id: I9fce3d3ede0289360824b8bc8e64b8426aa8451b