Update patch set 15
Patch Set 15:

> Tnx, testing...
 > 
 > Should I test all CE/PRO/Carrier here?

I've tested it on CE and PRO. CE passes all tests. On PRO it should work on active and passive nodes, although it complains about the permissions of the constants.yaml, which seems like a regression elsewhere. On the PRO standby node it also seems to complain about the ngcp-sems process which it thinks is running but that's not the case there so I'm a bit puzzled.

Testing on CARRIER would be appreciated!

Patch-set: 15
Label: Code-Review=0