Update patch set 2
Patch Set 2: Code-Review+2

+2

For this and some of the others I would probably give +1 to leave someone else to review or give advice in specific areas, or simply have time to notice the changes going on before being merged.

But since people will be on holidays, and we're experimenting with init system support after all and things are supposed to be shaky for a while, and the changes don't seem very harmful and can be tweaked or fixed or reverted, giving +2 for you to go ahead and merge if you want, so tests run overnight/over the weekend and you can continue later.

Patch-set: 2