TT#29356 Add Jessie based Dockerfile for Jenkins tap-tests
The image will be used to migrate all Jenkins tap-tests to Docker.
It will unblock the nodes "jenkins-slave*" migration to Debian stretch.

later we will be able to switch on "NGCP release specific Docker images"
(using Docker tags like we do for all other Docker images)
and then upgrade Docker image to Debian stretch in trunk
(while we still be able to keep jessie for some Git repos
in case of emergency).

Change-Id: I21fb5c088db3f2966b2e374647061059dda11329