Update patch set 1
Patch Set 1:

I have set +1 verified manually here as there is a bug in 
jenkins-debian-glue.
The wrapper perlcritic_tap consider the file is a perl file beucase it has "perl" into the file name for deleted file:

> file -b ../ngcp-panel/lib/NGCP/Panel.pm  
Perl5 module source, ASCII text

> file -b ../ngcp-panel/lib/NGCP/perlcriticrc 
cannot open `../ngcp-panel/lib/NGCP/perlcriticrc' (No such file or directory)

So the separate ticket with slow priority has been created
https://sipwise.my.workfront.com/task/view?ID=5a4e28dc001cafcb71e1e310fc15566f

and it is safe to merge this one as the issue related to the file move which is valid for this review only.

Patch-set: 1