TT#29920 Remove glusterfs port check, it fails very often nowadays
The glusterfs has ability to listen the next port in range if
the 'base-port' is occupied by some another process.

Which is happens very often on nightly tests in Debian stretch.
I have requested goss improvement to check the port from the range:
> https://github.com/aelsabbahy/goss/issues/311

So far we have to disable the check for the moment as it is unstable:
>> goss validate
> Port: tcp:24009: listening: doesn't match, expect: [true] found: [false]

While glusterfs is OK (but listening port 24010):

> root@sp1:~# netstat -anp | grep glusterfs
> tcp        0      0 0.0.0.0:24010           0.0.0.0:*               LISTEN      5412/glusterfsd
>
> root@sp1:~# gluster volume status ngcp
> Status of volume: ngcp
> Gluster process                             TCP Port  RDMA Port  Online  Pid
> ------------------------------------------------------------------------------
> Brick sp1:/var/lib/glusterfs/export         24010     0          Y       5412
> Brick sp2:/var/lib/glusterfs/export         24009     0          Y       5406

Change-Id: Idaf72caf4388be9b2bf5b0e98f1d4539d9438433
(cherry picked from commit f0f6422211d5f0e63ab537b1b5a08fa5f1ab4009)