Update patch set 3
Patch Set 3: Code-Review+2

Removed Manuel -2 here, IMHO it is not a blocker here too.
Also it is not related to this review at all IMHO, while maybe it worth to sync the list to have less questions...


Also I was thinking to create some metapackage with the list and operate with those metapackage in all the places, instead of hardcode the list everywhere.

Patch-set: 3
Reviewer: Alexander Lutay <1000001@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+2