TT#18774 Migrate NGCP from sysv to systemd
In systemd, getty is not spawned unconditionally, only when the
virtual consoled is enabled, so this process is not always running.

The config of systemd can be changed to start automatically getty
processes for the consoles and leave them waiting, but IMO changing
this just to satisfy the test is not worth it.

If in the past there have been problems with not being able to log
into the system because "getty" is not enabled maybe we can reconsider
this, or perhaps find another check that really checks if the full
login is possible (although it seems complicated), because just having
getty running is not enough.

Change-Id: Ibcfee42362457129f8d96c7c14e485de26e45083