TT#18774 Migrate NGCP from sysv to systemd
In systemd, motd is not running by default as a service.

The messages upon login on the vagrant VMs do show the custom MOTD
anyway, so whatever the mechanism that it's using, it seems to be
working fine.  I was doing some research and didn't find out the exact
mechanism.

In any case, either we change the test or we enable motd service under
systemd.

Change-Id: I5c082ef159fc1847de55925c1297006e120a38d5