Update patch set 1
Patch Set 1:

> > Isn't enough just to add libpam-systemd to make it work, while
 > keeping it socket-activated?
 > 
 > You can't have both works together, who will listen port 22 systemd
 > or sshd ?

I think that I didn't explain myself.

I didn't mean to keep it both as daemon and socket-activated at the same time.  I meant to keep socket-activation and while adding libpam-systemd to be installed.

AFAIK there are independent changes, so it's not needed to revert to daemon just to have libpam-systemd.

Patch-set: 1