Update patch set 1
Patch Set 1:

One thing to note though is that you're changing IPv6 to an IPv4 address, so maybe you can use ::1 instead?  Or both ::1 and 127.0.0.1?

Just an idea, but I expect that using IPv4 only for localhost doesn't create any problem for InfluxDB and its config on this system.

Patch-set: 1