Update patch set 1
Patch Set 1: Code-Review+1

Are we going to change the home directory for these daemons too? I'm not sure that's a good idea, as it diverges from upstream? Or is this just because whatever is doing the check does a realpath on the home dir?

Patch-set: 1
Reviewer: Guillem Jover <1000016@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+1