Update patch set 1
Patch Set 1:

Are we going to change the home directory for these daemons too? I'm not sure that's a good idea, as it diverges from upstream? Or is this just because whatever is doing the check does a realpath on the home dir?


I didn't change the home folder here, also I created the link /var/lib/mysql -> /ngcp-data/mysql  so I believe it can be even abandoned.

Patch-set: 1
Label: Code-Review=0