TT#42869 Remove kernel.sysrq check from system-tests
kernel.sysr=1 was set for Vagrant VMs only and missing
on the real hardware AND for old releases, also it is
going to be removed from installer as well, since it
was committed to debug Proxmox issues only (review pending).

No need to test it here. Also it will fix system-tests
for upgrade 2.8->...->mr6.5 which is currently failing:
> KernelParam: kernel.sysrq: value: doesn't match, expect: ["1"] found: ["438"]
(it happens as kernel.sysrq was not set for old releases).

Change-Id: Id3b449ab9a8b48df4404233c418263029e289359
(cherry picked from commit 1e6593d03fe6350ca54f712cdcb2841e013b49ea)