TT#46902 Remove tests of /var/cores
We have not been using /var/cores for a while, since starting to use
systemd-coredump around releases mr6.2 or mr6.3, so this file is now
unnecessary.

However, there was a script creating the file and this test ensuring that it
existed and no core dump files were present.

It's not particularly tied to TT#46902, but it got discovered while working on
it and the task also touches the area of "core dump" files.

Change-Id: If75fcfa28a75359033ab4efddec503da401e61bb