Update patch set 1
Patch Set 1:

> LGTM, so far double "command" check will be unnecessary if we are
 > going to use bind-mounts

If we use bind-mounts (*) we'd need to change also the filetype of /var/lib/systemd/coredump to be a directory.

(*) which I don't know if it's a good idea, for example in that case we would introduce dependencies and maybe ngcp-data volume could not be unmounted even in the case of emergencies.

Patch-set: 1
Label: Code-Review=0