TT#47373 Tune test for config.yml check in <=mr7.0
We cannot fully fix files permissions security issues before mr7.1
as new ngcpcfg with dependencies necessary there, see more details
in ticket TT#44808. We are going to announce the full feature support
in mr7.1+

In the same time some fixes/changes were backported to mr6.5 and
already released as a part of mr6.5.2, we need to tune system-tests
here as we might have different permissions for config.yml
on newly installed system and upgraded from previous release.
It is OK as for now and will be synchronised in mr7.1

Change-Id: I5aed15da183ae65070683fdfae310a45d2388573
(cherry picked from commit 6d55754e452d73115e8016909794f1886ef85910)