TT#57513 Adjust prosody check for lua version in Debian/buster
While we have lua5.1 on Debian/stretch, it's lua5.2 on Debian/buster
and lua5.3 is available as well in buster, so instead of hardcoding
the lua version 5.1, let's support whatever lua5.X version is used.

Fixes:

| not ok 34 - Command: pgrep -c -f '[l]ua5.1 /usr/bin/prosody': stdout: patterns not found: [1]

While at it, drop the unnecessary '[l]' approach, which
is a nice trick to avoid matching the grep process itself
within a `ps aux | grep ...`, but doesn't apply to
pgrep, thanks to Manuel Montecelo for spotting this leftover.

Change-Id: I3c62ddefbaf5a22f623b779426621ebaafec26b2