TT#68720 Mute awk cores on upgrade stretch->buster
The random awk crash happens which we cannot catch and trace.
It happens randomly and it produces no harm AFAIK.

We cannot report anything upstream here as we have to trace from core.
Also it is an old/stretch version of gawk which is crashed on stretch->buster
upgrade when gawk dependency libraries libc6, libgmp10, libsigsegv2
were fully/partially upgraded to buster already.

We can only mute it here as for now.
The mute is necessary in mr7.5* branched only
as mr7.5->mr8.0 is buster->buster upgrade already.

Change-Id: I24d15cb799f50cd5dbc0c5c99783d41ee32a780f