TT#72351 Move HAProxy port from 8080 to 5580
The port 8080 is commonly used to proxy HTTP traffic, and in our case we
were also using it for the voisniff XML-RPC interface, even though these
were supposed to run on different node roles, which was rather confusing
and made monitoring more difficult.

We move it to a port that ends in 80, and it's close to one of the other
prosody ports.

Change-Id: Ie6585b9f8450dbf3bda98126a8bd9e3b209f0908