TT#72351 Add missing dependencies
These are used by the tests themselves. We distinguish between core
and system dependencies for functionality we really want to check
which goes into Depends, and dependencies for stacks that are optional
and their missing presence will still emit an error due to the commands
not being available, these we add as Suggests.

Change-Id: I051dd388dab0a4a5c5f51f48dda54013298ba537