TT#66583 Update prosody version for mr8.2+ (to ensude upgrade went well)
We are not matchong to new prosody version 0.11.2-1 here as it can
be changed any time by Debian (as we migrated to upstream prosody version).

Also goss doesn't have support for regexps for 'versions' field,
so we cannot check something like:
>   prosody:
>     installed: true
>     versions:
>     - /0.11.2-.*/

After a long trying with goss 'matches/match-regexp/semver-constraint'
I decided to stop on the current version, where we inside old version is not available.

Change-Id: I626be7a23fd2b31d468f4cd9411530b24eb7216b