TT#71209 Enhance compare dbs tools
Use compare_dbs as a wrapper to check all the used schemes.

Add different formats of output to compare_dbs.pl - tap and human
readable one.
Process all the schemes in one run of compare_dbs.pl and do it in one db
connection.
If the tap formatter is used - output the result of all schemes in a
single file.

Change-Id: I2696680aa30b56658f130bd1cea116099c086753