TT#56053 Rebuild test files before running checks
Nowadays goss yml files are generated by ngcpcfg. But the content of
them depends on is node active or not. So if they are generated when
the node is in 'standby' state but then it's switched to 'active'
goss files still assume that the services are not run/enabled.

Rebuild them before running checks and commit with etckeeper. To do it
safely check if /etc directory is clean.

Change-Id: I19c3ef48f09c8d2e79613ef118adc362399a6e78