TT#56053 Make the node state a run-time variable
Nowadays goss yml files are generated by ngcpcfg. But their content
depends on whether the node is active or not, which can change
dynamically during the lifetime of the generated files from the
templates. Which will make the goss checks fail when the node state
changes.

Switch the node state into a run-time variable, and use go template
support to parametrize this into the goss checks.

Change-Id: I19c3ef48f09c8d2e79613ef118adc362399a6e78