TT#71209 Add exceptions to compare db tool
Sometimes there is divergence in schemes which are inevitable or
harmless.

So add exceptions' list which is easy to maintain. The format is the
following:
=================================
<element-type>/<schema-name>/<element-name>/<element-attribute>
Where:
  element-type - tables, columns, views, etc
  schema-name  - name of the schema
  element-name - name of the element (table, index, etc)
=================================

F.e.:
  views/ldap/ldap_entries/view_definition
For columns columns/<schema-name>/<table-name>_<column-name>
  columns/billing/table1_column1/is_nullable

Add 1st exception from TT#74728. The reason of this diff is that view
was created without schema name on sp1 but the replication statement was
written with schema name.

Change-Id: I29dd109cefb560fe89d92a430cd6522e4513f7d5