TT#71209 Add exceptions to compare db tool
Sometimes there is divergence in schemes which are inevitable or
harmless.

So add exceptions' list which is easy to maintain. The format is the
following:
type(table, trigger, etc)/schema name/element name/element
attribute(engine for table, column_type for column).

Add 1st exception from TT#74728. The reason of this diff is that view
was created without schema name on sp1 but the replication statement was
written with schema name.

Change-Id: I29dd109cefb560fe89d92a430cd6522e4513f7d5