TT#56053 Fix using of boolean SKIP* variables
The name of the variables SKIP_DNS_CHECK_TEST and
SKIP_STRICT_HOSTNAME_TEST so if they are 'true' the check should be
skipped. But in code it has direct branching, if the variable true -
run the test.

So add the negation to checks of these variables.

Change-Id: Iaf1e252dbb30cd2dc624db71742fa2cee8edb888
(cherry picked from commit 786bf567cea036c9c0f9d2181339564febb5b37b)