TT#84555 Check debug and messages files as optional
These files are created by rsyslog with default conf file. Usually it's
done after the deployment.sh's reboot.

But when NGCP is installed via installer package on existing debian
system there is no reboot between installation of rsyslog (as part of
NGCP) and ngcp-initial-configuration which modifies /etc/rsyslog.conf.
So in this case these files are not created.

So check the ownership and permission of these files only if they exist.

Change-Id: I33be13a4e78baee3926de8bd0135d655b89d5bb6