TT#71030 Fix influxdb checks conditionals for when it is present
InfluxDB is going to be running either when it is the current monitoring
backend, or when it is not but its data has not yet been migrated. We
want to check both cases so make sure the service is there and contains
data.

Change-Id: I45c71f41b6e2106b1ecf3a2665f669c920e59074
(cherry picked from commit 92029bad5f244c122df9f496778026cb3e4df9f2)