TT#106950 check that sshd is listening on localhost
Some of our scripts and templates rely on sshd listening on the
localhost addresses (e.g. the port test in monitrc). Normally this is
achieved either by having an ssh_ role assigned to the `lo` interface,
or by listening the addresses explicitly in the `listen_addresses` list.
Add an explicit test for this in case it is missing.

Change-Id: Iee409551e8237fbdec56fe33d1bb63d56d0af802