TT#106950 Revert "TT#106950 check that sshd is listening on localhost"
This reverts commit 4e164bf39fce05c5370331d4bfa3356ccfda0a81.

The change of enabling "::" as listening interface for sshd is only
applied in new installations, not on upgrades (neither as upgrade step
nor as cfg-schema script), so this test only works on newly installed
systems.

For systems installed before mr9.3 and upgraded, all systems fail at
this point of the tests.  Thus, all upgrade tests upgrading to mr9.3 are
failing at the moment because of this (and will keep failing for future
versions as well).

So unless and until there are plans to extend this change to systems
installed before mr9.3 by upgrading this configuration, we have to
revert this test.

Change-Id: Ic5e87651fab3daad9462c7135be81f37ef346950
(cherry picked from commit 816f4e5898596f3037186561d20b9eea372f695d)