TT#104381 Add missing tests of dirs in /ngcp-data/*, symlinks from /var/lib
For some reason some of the existing directories are tested while others
are not.

Add them, as some of them are created in the installer, so they will be
missing in upgrades installing through second root partition, and we
want to catch those cases before we start upgrading real systems.

Change-Id: If74db0fa91a16202d0c1297acc3d20abc3c09962