TT#105151 Reverse check to ensure that GRUB config does *not* contain net.ifnames=0
This has been applied before after changes in deployment-iso to not add
this parameter to GRUB config, but as it was reverted there it had to be
reverted here too.

We could just leave it untested, but in principle it's better to ensure
that we have what expect, either to ensure that it's present or absent,
depending to the actions in deployment-iso.

Change-Id: I695c9396a0f16987ee1146a153fffe1aafca8502