TT#156156 Switch to «-» as the word separator in CLI options
The prevalent convention in the project is to use «-» instead of «_» to
split words in options. Switch the options to that, and keep backwards
compatibility option alias that emits a deprecation warning while people
transition to the new names. The alias will be removed on the next LTS.

Clarify --help output.

Change-Id: Ic25ba7a152ce11c3be5106dc58d481add7ca02a8