TT#174750 Switch obsolete heartbeat-2 support to always disabled
Starting with mr10.5, the heartbeat-2 support is now obsolete and not
available anymore. Switch all heartbeat-2 service and process checks
to expect them to be disabled. Remove the UDP check, as we have now
removed the RPC port start limitation, so we might end up checking for
an unrelated colliding port.

Change-Id: Ibacaf2d55e45b3f86a5218b92204d180e934e1e2