TT#177600 Move SSL certificates under /etc/ngcp-config/shared-files/ssl
These are shared files across all PRO/Carrier nodes, but at the time
they were introduced we did not have the concept of a shared-files
subdirectory. All other such shared files have been moved over time to
that subdirectory, but this was still remaining behind. Let's move it
now for consistency, and so that its purpose is more clear.

Change-Id: I45c89245d36706dba5ed1fb47a8bf9124c390dae