MT#33035 Make user.log optional
It's an often error after upgrade that this file is missing.
It happens because right after reboot there is no such file and it's not
created by logrotate because we use missingok, notifempty option for
it.

So make it optional.

Change-Id: I4ee18e8ba5e93d19e4b80fd3fcf8197d144f1ea3
(cherry picked from commit 790ca5ce32991a5ce71c25638920478e69cc5ab2)