MT#56187 Fix rate-o-mat service check
The rate-o-mat service can run in active-active or in active-standby
modes. If the service is enabled and on the proxy nodes, the former
mode means the service will be enabled and running on both peers, the
latter will mean it's in the traditional HA mode and thus only running
on the active node.

Change-Id: I020c8a00706135ed5d432bf8b1b8874cf1b2f532