MT#58452 Do not expect NGINX to listen on web_int for localhost
There's currently a divergence between CE and PRO, where web_int is not
setup by default as a role for the loopback interface on CE systems. We
should thus for now not expect NGINX to be listening there, for the
NGCP Panel admin and csc ports.

In the future we should probably unify this behavior and make CE behave
the same as a PRO, because this seems like a gratuitous divergence.

Fixes: commit d0d8c1eb102023fd5968d304ffa2667b18b07050
Change-Id: Ib65b9dcf94a34b416d59aad93e19d88cf5a6469c