MT#61304 Handle conflicting Prosody and HAProxy ports
These share the same ports depending on the role of the current node, so
to avoid emitting the same duplicate port entries we should turn these
into cascading if/elif.

This is being detected by new goss versions from Debian trixie as
errors, and causing the entire ngcp-system-tests run to fail.

Change-Id: I2898f623e87867a03ef6cfb728a90631f53ffae7