MT#62935 Update sysctl kernel.core_pattern for systemd 252.38-1~deb12u1 security update
The systemd package got a security update in Debian bookworm, which
changed the systemd-coredump kernel sysctl core_pattern value (by
appending « %d»). This is part of the fix for CVE-2025-4598.

This has caused ngcp-system-tests to fail to match the new pattern,
so we need to adapt it for all currently support Debian bookworm
releases.

Ref: https://security-tracker.debian.org/tracker/CVE-2025-4598
Ref: https://github.com/systemd/systemd-stable/commit/2eb46dce078334805c547cbcf5e6462cf9d2f9f0
Change-Id: I531f197e47094321d688d425fb7f577b42fd7391