MT#62935 Update sysctl kernel.core_pattern for systemd 257.6-1 security update
The systemd package got a security update in Debian trixie, which
changed the systemd-coredump kernel sysctl core_pattern value (by
appending « %d %F»). This is part of the fix for CVE-2025-4598.

This has caused ngcp-system-tests to fail to match the new pattern,
so we need to adapt it for all currently supported Debian trixie
releases.

In addition for Debian trixie, a new enough Linux kernel is required
to benefit from the full security fix.

Ref: https://security-tracker.debian.org/tracker/CVE-2025-4598
Ref: https://github.com/systemd/systemd-stable/commit/2eb46dce078334805c547cbcf5e6462cf9d2f9f0
Change-Id: I89e6ff1d084403e6ae4b4eca6f5606b0d2417c01