MT#62899 Update copyright holder and license from Artistic + GPL-1+ to GPL-3+
In commit ddd3277692a6711a714d4e054451387c9abfe121 we replaced the
previous code using the imported Perl code from Jozef Kutej, to
an implementation based on the goss tool.

The Artistic + GPL-1+ license was due to the imported code, but that is
no longer applicable, so we can safely remove the copyright statement
for Jozef Kutej and switch the license to match what other GPL projects
in NGCP use.

Change-Id: Ia8ac16993b03758a9f23b5021e4236c9a32e5df0